the store and reducer have been modified, and the redux toolkit dependency has been reinstalled
