+++ src/redux/reducer/index.js
@@ -1,0 +1,12 @@
+const initialState = {
+    instruments: [],
+    allInstruments:[]
+}
+
+export default function rootReducer(state = initialState, action){
+    switch(action.type){
+        
+        default: 
+        return state
+    }
+}
