--- src/redux/reducers/index.js
@@ -1,19 +1,0 @@
-const initialState = {
-    
-    instruments: [],
-    allInstruments:[]
- 
-}
-
-export default function rootReducer(state = initialState,action){
-    switch(action.type){
-        case 'GET_ALL_INSTRUMENTS':
-            return {
-                ...state,
-                allInstruments: action.payload,
-                instruments: [...action.payload]
-            }
-            default: 
-            return state
-    }
-}
