+++ src/redux/store.js
@@ -1,8 +1,16 @@
+import { compose, applyMiddleware } from "redux";
 import { configureStore } from "@reduxjs/toolkit";
-import isLoadingReducer from "./reducers/isLoadingSlice";
+import rootReducer from "./reducer";
+import thunk from "redux-thunk";
 
-export default configureStore({
-    reducer: {
-        isloading: isLoadingReducer,
-    }
-});
+const composeEnhancers =
+    (typeof window !== 'undefined' &&
+        window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+    compose;
+
+const store = configureStore(
+    {reducer: rootReducer},
+    composeEnhancers(applyMiddleware(thunk)),
+);
+
+export default store;
